Deploy_2/requirements.txt
streamlit
sklearn
pandas
joblib
numpy
plotly.express
xgboost